+++ packages/miniapp-render/src/constants.js
@@ -1,3 +1,6 @@
+// eslint-disable-next-line import/no-extraneous-dependencies
+import { isBaiduSmartProgram, isKuaiShouMiniProgram } from 'universal-env';
+
 export const NATIVE_EVENTS_LIST = [
   'onBack',
   'onKeyboardHeight',
@@ -28,11 +31,15 @@ export const BODY_NODE_ID = 'e-body';
 
 export const INDEX_PAGE = 'index-page';
 
-export const STATIC_COMPONENTS = new Set(['view', 'text', 'image']); // With no events components
+/**
+ * In Baidu SmartProgram and KuaiShou MiniProgram, view and h-element template are modified to support flex, so pure, static and catch elements are omitted
+ */
 
-export const PURE_COMPONENTS = new Set(['view', 'h-element']); // With no events or props && equal to TOUCH_COMPONENTS
+export const STATIC_COMPONENTS = new Set(isBaiduSmartProgram || isKuaiShouMiniProgram ? ['text', 'image'] : ['view', 'text', 'image']); // With no events components
 
-export const CATCH_COMPONENTS = new Set(['view', 'h-element']); // With catchTouchMove events
+export const PURE_COMPONENTS = new Set(isBaiduSmartProgram || isKuaiShouMiniProgram ? [] : ['view', 'h-element']); // With no events or props && equal to TOUCH_COMPONENTS
+
+export const CATCH_COMPONENTS = new Set(isBaiduSmartProgram || isKuaiShouMiniProgram ? [] : ['view', 'h-element']); // With catchTouchMove events
 
 export const APPEAR_COMPONENT = 'view'; // Without appear event components