+++ packages/jsx-compiler/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## [0.4.35] - 2021-08-27
+
+- Return renamed index in list
+
 ## [0.4.34] - 2021-08-24
 
 - Compatible with old version of jsx2mp-runtime, which not support list key